add postgresql package and refactor login
