+++ package.json
@@ -45,6 +45,7 @@
 		"express": "^4.18.2",
 		"mssql": "^9.0.1",
 		"mysql2": "^2.3.3",
+		"postgres": "^3.3.2",
 		"svelte-dialogs": "^1.2.2",
 		"sveltestrap": "^5.9.0"
 	}
